Refactored viz module with composables, services, etc
