+++ lib/osh-js/source/core/ui/view/map/CesiumView.js
@@ -487,7 +487,7 @@ class CesiumView extends MapView {
                 // description: undefined,
                 // position: undefined,
                 image: properties.icon,
-                scaleByDistance: new NearFarScalar(1000, 0.5, 100000, 0.3),
+                scaleByDistance: new NearFarScalar(1000, 1.0, 10e6, 0.0),
                 // alignedAxis: (this.viewer.camera.pitch < -Math.PI / 4) && properties.allowBillboardRotation ? Cartesian3.UNIT_Z : Cartesian3.ZERO, // Z means rotation is from north
                 // rotation: rot,
                 // horizontalOrigin: HorizontalOrigin.LEFT,
