+++ src/modules/visualization/visualizations/chart/Config.vue
@@ -2,8 +2,8 @@
 import { useVizWizStore } from '@/stores/vizwizstore';
 import { ref, computed, reactive, watch, ReactiveEffect, onMounted } from 'vue';
 import DataSourcePicker from '../../wizard/components/DataSourcePicker.vue';
-import { useComponentValidation } from '../../../../components/menus/visualization-wizard/shared/helpers';
 import { VisualizationComponentEmits } from '../../registry/VisualizationRegistry';
+import { useComponentValidation } from '../../wizard/composables/useComponentValidation';
 
 // Retrieve datastreams
 const vizwizStore = useVizWizStore()
