+++ src/modules/visualization/visualizations/lob/Config.vue
@@ -3,7 +3,7 @@ import { useVizWizStore } from '@/stores/vizwizstore';
 import { ref, computed, reactive, watch, ReactiveEffect, onMounted } from 'vue';
 import DataSourcePicker from '../../wizard/components/DataSourcePicker.vue';
 import { VisualizationComponentEmits } from '../../registry/VisualizationRegistry';
-import { useComponentValidation } from '../../../../components/menus/visualization-wizard/shared/helpers';
+import { useComponentValidation } from '../../wizard/composables/useComponentValidation';
 
 
 // Retrieve datastreams
