+++ src/modules/visualization/visualizations/video/Config.vue
@@ -4,7 +4,7 @@ import { computed, reactive, watch, onMounted, ref } from 'vue';
 import DataSourcePicker from '../../wizard/components/DataSourcePicker.vue';
 import ControlStreamPicker from '../../wizard/components/ControlStreamPicker.vue';
 import { VisualizationComponentEmits } from '../../registry/VisualizationRegistry';
-import { useComponentValidation } from '../../../../components/menus/visualization-wizard/shared/helpers';
+import { useComponentValidation } from '../../wizard/composables/useComponentValidation';
 
 
 // Retrieve datastreams
