+++ src/modules/visualization/visualizations/video/Customize.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { computed, ref } from 'vue';
-import { useComponentValidation } from '../../../../components/menus/visualization-wizard/shared/helpers';
 import { VisualizationComponentEmits } from '../../registry/VisualizationRegistry';
 import NameControl from '@/modules/visualization/wizard/customizations/NameControl.vue';
 import VideoOptions from '@/modules/visualization/wizard/customizations/VideoOptions.vue'
+import { useComponentValidation } from '../../wizard/composables/useComponentValidation';
 
 // Validation: Name cannot be empty
 const emit = defineEmits<VisualizationComponentEmits>()
