+++ src/modules/visualization/visualizations/video/Video.vue
@@ -3,14 +3,14 @@ import { computed, onMounted, ref, toRaw } from 'vue';
 import { randomUUID } from 'osh-js/source/core/utils/Utils.js';
 import VideoDataLayer from 'osh-js/source/core/ui/layer/VideoDataLayer.js';
 import { OSHControlStream, OSHVisualization } from '@/lib/OSHConnectDataStructs';
-import MJPEGView from 'osh-js/source/core/ui/view/video/MjpegView.js';
 import VideoView from 'osh-js/source/core/ui/view/video/VideoView.js';
 import SweApi from 'osh-js/source/core/datasource/sweapi/SweApi.datasource.js';
 import PTZControl from './PTZControl.vue'
 import { useControlStreamStore } from "@/stores/controlstreamstore";
 import { fetchControlStreamSchema } from "@/lib/ControlstreamUtils";
-import { createDatasource, useVisualizationCleanup } from '../../../../components/menus/visualization-wizard/shared/helpers';
+import { createDatasource } from '@/modules/visualization/services/datasource.service';
 import { ISweApiDataSourceProperties, IVideoLayerProperties, IVideoViewProperties } from '@/lib/VisualizationHelpers';
+import { useVisualizationCleanup } from '../../components/composables/useVisualizationCleanup';
 
 const props = defineProps<{
   visualization: OSHVisualization,
