+++ src/modules/visualization/wizard/customizations/NameControl.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import { watch, ref, onMounted, computed } from 'vue';
 import { useVizWizStore } from '@/stores/vizwizstore';
-import { generateVizName, useComponentValidation } from '../../../../components/menus/visualization-wizard/shared/helpers';
 import { VisualizationComponentEmits } from '../../registry/VisualizationRegistry';
+import { useComponentValidation } from '../composables/useComponentValidation';
 
 const props = defineProps<{
   role?: string;  // Role to use with generateVizName function
@@ -37,6 +37,23 @@ const valid = computed(() => {
   return !!name.value
 })
 useComponentValidation(valid, emit)
+
+/**
+ * Generates a default visualization name based on selected viz type and a given role's datastream name
+ * 
+ * @param role
+ * @returns 
+ */
+function generateVizName(role: string) {
+  // Find datastream ID of desired role
+  const dsId = vwStore.dsConfig[role].dsId
+
+  for (const ds of vwStore.datastreams) {
+    if (ds.id === dsId) return `${vwStore.visualizationType}: ${ds.name}`
+  }
+
+  return `New ${vwStore.visualizationType}`
+}
 
 </script>
 <template>
